requirements.txt
torch>=2.1.1
torchvision>=0.16.1
torchaudio>=2.1.1
numpy>=1.23.5
timm>=0.9.2
mlflow>=2.9.1
opencv-python>=4.8.1
Pillow>=11.3.0
matplotlib>=3.10.3
scipy>=1.11.4
scikit-learn>=1.6.1
einops>=0.7.0